requirements.txt
environs==14.1.1
python-telegram-bot==13.2
requests==2.32.3